+++ packages/db/package.json
@@ -29,7 +29,7 @@
     "build": "bun run clean && tsc -b",
     "clean": "rm -rf ./dist ./tsconfig.tsbuildinfo",
     "dev": "tsc -b -w",
-    "dev:db": "turso dev --db-file devv2.db",
+    "dev:db": "turso dev --db-file dev.db",
     "db:generate": "NODE_ENV=production drizzle-kit generate  --config=drizzle-production.config.ts",
     "db:generate:dev": "NODE_ENV=development drizzle-kit generate --config=drizzle-development.config.ts",
     "db:migrate": "NODE_ENV=production drizzle-kit migrate --config=drizzle-production.config.ts",
